completed basic logic of customer

+++ internal/application/order/service.go
@@ -19,7 +19,7 @@ func NewService(repo Repository, paymentPort ports.PaymentPort) *Service {
 	}
 }
 
-func (s *Service) CreateOrder(customerID uint64, items []domain.OrderItem) (*domain.Order, error) {
+func (s *Service) ProcessOrder(customerID uint64, items []domain.OrderItem) error {
 	order := &domain.Order{
 		CustomerID: customerID,
 		Items:      items,
@@ -28,30 +28,20 @@ func (s *Service) CreateOrder(customerID uint64, items []domain.OrderItem) (*dom
 
 	err := s.repo.Save(order)
 	if err != nil {
-		return nil, err
+		return err
 	}
 
 	// pay for the order
 	totalPrice := order.TotalPrice()
 	res, err := s.payment.Charge(customerID, order.ID, totalPrice)
 
-	// case 1: payment error
-	if err != nil {
-		order.Status = "unpaid"
-		s.repo.Update(order)
-		return nil, errors.New("failed to process payment")
-	}
-	// case 2: successfully paid
-	if res.Status == "success" {
+	// successfully paid
+	if err == nil && res.Status == "success" {
 		s.repo.Delete(order.ID)
-		return order, nil
+		return nil
 	}
 
 	order.Status = "unpaid"
 	s.repo.Update(order)
-	return nil, errors.New(res.Message)
-}
-
-func (s *Service) GetUnpaidOrders(userID uint64) ([]domain.Order, error) {
-	return s.repo.FindUnpaidByUser(userID)
+	return errors.New(res.Message)
 }
